requirements.txt
Flask==3.0.0
pycryptodome==3.19.0
werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
